requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pytest>=7.4.0
pytest-html>=4.0.0